PWN-6775 - Resolve PR comments
